LayoutTests/fast/frames/sandboxed-iframe-attribute-parsing-03-expected.txt

CONSOLE MESSAGE: line 46: Error while parsing the 'sandbox' attribute: 'För', 'var', 'vers,', 'jag', 'gör,', 'Lovar', 'du', 'en', 'kyss', 'mig', 'giva;', 'Arket', 'fullt', 'jag', 'borde', 'skriva,', 'Mindre', 'har', 'jag', 'skrivit', 'för.', 'Men', 'man', 'måste', 'hålla', 'måtta,', 'Jag', 'med', 'vers,', 'med', 'kyssar', 'du.', 'Låt', 'mig', 'räkna:', 'Där', 'är', 'sju!', 'En', 'därtill', 'det', 'gör', 'mig', 'åtta.', 'Numro', 'åtta', 'är', 'fatal,', 'Greklands', 'sångmör', 'voro', 'nio,', 'Och', 'en', 'svensk', 'därtill', 'gör', 'tio.', '—', 'Elva', 'var', 'apostelns', 'tal,', 'Ty', 'jag', 'räknar', 'icke', 'Judas,', 'Honom,', 'som', 'i', 'vänners', 'lag', 'Kysste', 'falskt;', 'det', 'gör', 'ej', 'jag,', 'Helst', 'när', 'vackra', 'läppar', 'bjudas.', 'Huru', 'står', 'min', 'räkning', 'här?', 'Aderton;', 'det', 'är', 'dock', 'något.', 'Nitton', '—', 'rimmet', 'gör', 'besvär,', 'Därföre', 'jag', 'fyller', 'tjoget.', 'Strofen', 'är', 'ej', 'full', 'som', 'jag,', 'In', 'i', 'hamnen', 'vill', 'jag', 'styra,', 'Därföre', 'till', 'godo', 'tag', 'Denna', 'gång', 'med', 'tjugofyra.', ''Kyssarna'', '('The', 'kisses'),', 'Esaias', 'Tegnér,', '1782-1846', 'int', 'main(void)', '{', 'return', '0;', '}' are invalid sandbox flags.
CONSOLE MESSAGE: line 7: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
Ridiculously long, invalid text (well, for these purposes at least) with non-ASCII characters surrounding attribute value


PASS
